Make EntityEntryGraphNode.TState mutable
Fixes #18823
